Fix warning: implicit conversion changes signedness: 'uint32_t' (aka 'unsigned int') to 'int' [-Wsign-conversion]
visp/modules/core/src/math/random-generator/vpUniRand.cpp:166:29: warning: implicit conversion changes signedness: 'uint32_t' (aka 'unsigned int') to 'int' [-Wsign-conversion]
  166 |   return boundedRand(b - a) + a;
      |   ~~~~~~ ~~~~~~~~~~~~~~~~~~~^~~
visp/modules/core/src/math/random-generator/vpUniRand.cpp:166:31: warning: implicit conversion changes signedness: 'int' to 'uint32_t' (aka 'unsigned int') [-Wsign-conversion]
  166 |   return boundedRand(b - a) + a;
      |                             ~ ^
visp/modules/core/src/math/random-generator/vpUniRand.cpp:166:24: warning: implicit conversion changes signedness: 'int' to 'uint32_t' (aka 'unsigned int') [-Wsign-conversion]
  166 |   return boundedRand(b - a) + a;
      |          ~~~~~~~~~~~ ~~^~~

+++ modules/core/src/math/random-generator/vpUniRand.cpp
@@ -1,6 +1,6 @@
 /*
  * ViSP, open source Visual Servoing Platform software.
- * Copyright (C) 2005 - 2024 by Inria. All rights reserved.
+ * Copyright (C) 2005 - 2026 by Inria. All rights reserved.
  *
  * This software is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -71,7 +71,7 @@
 BEGIN_VISP_NAMESPACE
 vpUniRand::vpUniRand()
   : m_maxInvDbl(1.0 / static_cast<double>(UINT32_MAX)), m_maxInvFlt(1.0f / static_cast<float>(UINT32_MAX)), m_rng()
-{ }
+{}
 
 /*!
   Create a pseudorandom number generator with uniform distribution.
@@ -163,7 +163,7 @@ int vpUniRand::uniform(int a, int b)
   if (a == b) {
     return a;
   }
-  return boundedRand(b - a) + a;
+  return static_cast<int>(boundedRand(static_cast<uint32_t>(b - a))) + a;
 }
 
 /*!
